requirements.txt
# /workspace/requirements.txt (Simplified Constraints)

# --- Core ML/Vision/LLM Libraries ---
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0 # Keep a reasonable minimum, but no upper bound for now
accelerate>=0.25.0
bitsandbytes>=0.41.3
timm>=1.0 # REQUIRED for DETR models

# --- LangChain & RAG Components ---
# Let pip find compatible 0.1.x versions or slightly newer
langchain>=0.1.5,<0.2.0
langchain-community>=0.0.17,<0.1.0
langchain-core>=0.1.18,<0.2.0 # Ensure a base version but allow compatible upgrades
langchain-huggingface>=0.0.1 # Needs langchain-core >= 0.1.52 usually
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rank_bm25>=0.2.2
tiktoken>=0.5.2
# REMOVE explicit langsmith pin - let langchain handle it
# REMOVE explicit huggingface_hub pin - let transformers/others handle it

# --- PDF & Image Handling ---
pdf2image>=1.16.3
Pillow>=9.4.0
pymupdf>=1.23.7
opencv-python>=4.8.0

# --- OCR ---
pytesseract>=0.3.10

# --- Table Extraction ---
img2table>=1.2.7
tabulate>=0.9.0
camelot-py[cv]>=0.11.0 # Keep this pinned as 0.11.0 was specified before

# --- UI ---
gradio>=4.10.0,<4.11.0 # Keep Gradio reasonably pinned
fastapi==0.104.1 # Pin FastAPI version to avoid Pydantic/Starlette conflict

# --- Dev & Utility ---
jupyterlab>=4.0.0 # Optional
ipykernel>=6.25.0 # Optional
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0 # Optional
joblib>=1.3.2
protobuf<4,>=3.20
scipy
typing-inspect>=0.8.0
typing_extensions>=4.5.0
pydantic<3,>=1 # Keep pydantic constraint